camera: added preferred_stride control
This change fixes OP-705.

This control makes it possible to change the stride length. This
affects sensors like IMX390. As an example IMX390 has a resolution of
1936x1100. Without altering the preferred stride raw captures will
have incorrect alignment. This is seen as jagged rows. By setting the
preferred stride for IMX390 to 4096 each row will occupy 2048
bytes. There will be black padding bytes but these can easily be
removed.

Change-Id: I67547666b816df95feeefb8d4dd19218f9d11ab8